refactor, seperate camera, camera_settings and camera_holder
